Code refactor according to review.
