+++ src/frontend/org/voltcore/utils/CoreUtils.java
@@ -59,6 +59,7 @@ import java.util.concurrent.TimeoutException;
 import java.util.concurrent.atomic.AtomicLong;
 import java.util.concurrent.locks.ReentrantLock;
 
+import org.voltcore.logging.Level;
 import org.voltcore.logging.VoltLogger;
 import org.voltcore.network.ReverseDNSCache;
 
@@ -1199,6 +1200,60 @@ public class CoreUtils {
         }
         catch (Exception e) {
             log.fatal("Unable to list ports in use at this time.");
+        }
+    }
+
+    /**
+     * Print logs surrounded by stars
+     * @param vLogger
+     * @param msg
+     * @param level
+     */
+    public static void PrintGoodLookingLog(VoltLogger vLogger, String msg, Level level)
+    {
+        if (vLogger == null || msg == null || level == Level.OFF) { return; }
+
+        // The border of the star-surrounded message
+        //      **********
+        //      * msg... *
+        //      **********
+        // Improvement may be needed to handle multi-line messages
+        String stars = String.join("", Collections.nCopies(msg.length() + 4, "*"));
+        String new_msg = "* " + msg + " *";
+
+        switch (level) {
+            case DEBUG:
+                vLogger.debug(stars);
+                vLogger.debug(new_msg);
+                vLogger.debug(stars);
+                break;
+            case WARN:
+                vLogger.warn(stars);
+                vLogger.warn(new_msg);
+                vLogger.warn(stars);
+                break;
+            case ERROR:
+                vLogger.error(stars);
+                vLogger.error(new_msg);
+                vLogger.error(stars);
+                break;
+            case FATAL:
+                vLogger.fatal(stars);
+                vLogger.fatal(new_msg);
+                vLogger.fatal(stars);
+                break;
+            case INFO:
+                vLogger.info(stars);
+                vLogger.info(new_msg);
+                vLogger.info(stars);
+                break;
+            case TRACE:
+                vLogger.trace(stars);
+                vLogger.trace(new_msg);
+                vLogger.trace(stars);
+                break;
+            default:
+                break;
         }
     }
