+++ src/frontend/org/voltcore/utils/ShutdownHooks.java
@@ -25,7 +25,6 @@ import java.util.TreeMap;
 
 import org.voltcore.logging.Level;
 import org.voltcore.logging.VoltLogger;
-import org.voltdb.utils.LoggerHelper;
 
 /**
  * A hacky singleton which provides simple coordinated registration of ShutdownHooks within Volt.
@@ -133,7 +132,7 @@ public class ShutdownHooks
         if (m_iAmAServer && !m_crashing && ShutdownHooks.m_crashMessage) {
             VoltLogger voltLogger = new VoltLogger("CONSOLE");
             String msg = "The VoltDB server will shut down due to a control-C or other JVM exit.";
-            LoggerHelper.PrintGoodLookingLog(voltLogger, msg, Level.WARN);
+            CoreUtils.PrintGoodLookingLog(voltLogger, msg, Level.WARN);
         }
         for (Entry<Integer, List<ShutdownTask>> tasks : m_shutdownTasks.entrySet()) {
             for (ShutdownTask task : tasks.getValue()) {
