+++ src/frontend/org/voltdb/sysprocs/Shutdown.java
@@ -23,6 +23,7 @@ import java.util.concurrent.atomic.AtomicBoolean;
 
 import org.voltcore.logging.Level;
 import org.voltcore.logging.VoltLogger;
+import org.voltcore.utils.CoreUtils;
 import org.voltdb.DependencyPair;
 import org.voltdb.ParameterSet;
 import org.voltdb.ProcInfo;
@@ -33,7 +34,6 @@ import org.voltdb.VoltTable;
 import org.voltdb.VoltTable.ColumnInfo;
 import org.voltdb.VoltType;
 import org.voltdb.dtxn.DtxnConstants;
-import org.voltdb.utils.LoggerHelper;
 
 /**
  * A wholly improper shutdown. No promise is given to return a result to a client,
@@ -58,7 +58,7 @@ public class Shutdown extends VoltSystemProcedure {
             } catch (InterruptedException e) {}
             VoltLogger voltLogger = new VoltLogger("HOST");
             String msg = "VoltDB shutting down as requested by @Shutdown command.";
-            LoggerHelper.PrintGoodLookingLog(voltLogger, msg, Level.WARN);
+            CoreUtils.PrintGoodLookingLog(voltLogger, msg, Level.WARN);
             System.exit(0);
         }
     };
@@ -85,7 +85,7 @@ public class Shutdown extends VoltSystemProcedure {
                 m_failsafe.start();
                 VoltLogger voltLogger = new VoltLogger("HOST");
                 String msg = "VoltDB shutdown operation requested and in progress. Cluster will terminate shortly.";
-                LoggerHelper.PrintGoodLookingLog(voltLogger, msg, Level.WARN);
+                CoreUtils.PrintGoodLookingLog(voltLogger, msg, Level.WARN);
             }
             VoltTable rslt = new VoltTable(new ColumnInfo[] { new ColumnInfo("HA", VoltType.STRING) });
             return new DependencyPair.TableDependencyPair(DEP_shutdownSync, rslt);
@@ -109,7 +109,7 @@ public class Shutdown extends VoltSystemProcedure {
                     if (die) {
                         VoltLogger voltLogger = new VoltLogger("HOST");
                         String msg = "VoltDB shutting down as requested by @Shutdown command.";
-                        LoggerHelper.PrintGoodLookingLog(voltLogger, msg, Level.WARN);
+                        CoreUtils.PrintGoodLookingLog(voltLogger, msg, Level.WARN);
                         System.exit(0);
                     }
                     else {
